backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
alembic
pydantic
pydantic-settings
pyjwt
passlib[bcrypt]
email-validator
